requirements.txt
# Core Data Lake & Processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0
delta-lake>=0.15.0
azure-storage-blob>=12.17.0
boto3>=1.26.0

# Meta Layer & Semantic Modeling
pydantic>=2.0.0
rdflib>=7.0.0
owlready2>=0.46
networkx>=3.0.0

# AI Agents Framework
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.0.3
crewai>=0.28.0
openai>=1.6.0
anthropic>=0.8.0

# Database & Query
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
sqlite3

# Web & API
fastapi>=0.104.0
uvicorn>=0.24.0
streamlit>=1.28.0

# Clinical Data Standards
cdisc-library>=0.3.0
pandas-profiling>=3.6.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
loguru>=0.7.0
rich>=13.6.0
typer>=0.9.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
